+++ packages/terra-application/tests/jest/application-base/ApplicationBase.test.jsx
@@ -36,4 +36,12 @@ describe('ApplicationBase', () => {
     ));
     expect(wrapper).toMatchSnapshot();
   });
+  it('should render without scroll', () => {
+    const wrapper = shallow((
+      <ApplicationBase noScroll>
+        <div>content</div>
+      </ApplicationBase>
+    ));
+    expect(wrapper).toMatchSnapshot();
+  });
 });