+++ packages/terra-application/tests/jest/application-status-overlay/ApplicationStatusOverlayProvider.test.jsx
@@ -149,5 +149,13 @@ describe('ApplicationStatusOverlayProvider', () => {
 
       React.useState = reactUseState;
     });
+
+    it('should render without scroll', () => {
+      const wrapper = shallow((
+        <ApplicationStatusOverlayProvider noScroll />
+      ));
+
+      expect(wrapper).toMatchSnapshot();
+    });
   });
 });